requirements.txt
google-generativeai
SpeechRecognition
PyAudio
PyQt5
pyttsx3